requirements.txt
browserid==1.0
Django==1.8.4
django-browserid==1.0.0
django-dajax==0.9.2
django-dajaxice==0.7
django-extras==0.3
django-materialize-css==0.0.1
django-realtime==1.1
extras==0.0.3
requests==2.7.0
six==1.9.0
